requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pyyaml==6.0.1
transformers>=4.48.0
torch>=2.1.0
numpy
pytest>=7.0.0
